cleanup testing folder
