--- testing/EmailTesting/EmailTesting/README.md
@@ -1,3 +1,0 @@
-send using email_sender.py
-
-check to see if mail is sent using spiders/main.py
