+++ tests/EmailTesting/EmailTesting/pipelines.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+
+# Define your item pipelines here
+#
+# Don't forget to add your pipeline to the ITEM_PIPELINES setting
+# See: https://docs.scrapy.org/en/latest/topics/item-pipeline.html
+
+
+class EmailtestingPipeline(object):
+    def process_item(self, item, spider):
+        return item
