+++ tests/README.md
@@ -1,0 +1,1 @@
+set up Selenium
